refactor(compact_sideboard): split CompactSideboardPanel into mixin package
Move widgets/panels/compact_sideboard_panel.py into a package with
frame.py for UI construction and handlers.py for the public display
setters, the on_toggle_play_draw callback, and the list populator.
