+++ widgets/panels/compact_sideboard_panel/__init__.py
@@ -1,0 +1,7 @@
+"""Compact sideboard guide panel package."""
+
+from __future__ import annotations
+
+from widgets.panels.compact_sideboard_panel.frame import CompactSideboardPanel
+
+__all__ = ["CompactSideboardPanel"]
